F added configuration property to RedisFactory F added redisConfigurationFactory that allows to store a configuration for later client creation F redisConfigurationFactory has a list of static method replacing old RedisConfiguration initializers, this is essential in order allow discovery on  redis.use(...) R changed type on redisStorage
